Move hash test to test_hash.py
